requirements.txt
customtkinter
rembg[cpu]
pillow
opencv-python
tkinterdnd2-universal
onnxruntime